Feat/support local task reload

+++ dfget/core/uploader/peer_server.go
@@ -19,13 +19,16 @@ package uploader
 import (
 	"bytes"
 	"context"
+	"encoding/base64"
 	"encoding/binary"
 	"encoding/json"
 	"fmt"
 	"io"
+	"io/ioutil"
 	"net"
 	"net/http"
 	"os"
+	"path/filepath"
 	"strconv"
 	"strings"
 	"sync"
@@ -38,6 +41,9 @@ import (
 	"github.com/dragonflyoss/Dragonfly/pkg/limitreader"
 	"github.com/dragonflyoss/Dragonfly/pkg/ratelimiter"
 	"github.com/dragonflyoss/Dragonfly/version"
+	apitypes "github.com/dragonflyoss/Dragonfly/apis/types"
+	"github.com/dragonflyoss/Dragonfly/pkg/constants"
+	"github.com/dragonflyoss/Dragonfly/dfget/types"
 
 	"github.com/gorilla/mux"
 	"github.com/sirupsen/logrus"
@@ -53,6 +59,14 @@ func newPeerServer(cfg *config.Config, port int) *peerServer {
 		api:      api.NewSupernodeAPI(),
 	}
 
+	s.syncTaskContainer = &taskContainer{
+		syncTaskMap: sync.Map{},
+		ps: 	s,
+	}
+
+	logrus.Infof("peer server config: %v, supernodes: %v", cfg, cfg.Nodes)
+
+	s.reload()
 	r := s.initRouter()
 	s.Server = &http.Server{
 		Addr:    net.JoinHostPort(s.host, strconv.Itoa(port)),
@@ -69,7 +83,7 @@ func newPeerServer(cfg *config.Config, port int) *peerServer {
 type peerServer struct {
 	cfg *config.Config
 
-	// finished indicates whether the peer server is shutdown
+	// Finished indicates whether the peer server is shutdown
 	finished chan struct{}
 
 	// server related fields
@@ -83,21 +97,69 @@ type peerServer struct {
 	// totalLimitRate is the total network bandwidth shared by tasks on the same host
 	totalLimitRate int
 
-	// syncTaskMap stores the meta name of tasks on the host
+	// syncTaskContainer stores the meta name of tasks on the host
+	syncTaskContainer *taskContainer
+}
+
+type taskContainer struct {
 	syncTaskMap sync.Map
+	ps		*peerServer
+}
+
+func (t *taskContainer) Store(taskFileName string, tc *taskConfig) {
+	t.syncTaskMap.Store(taskFileName, tc)
+	t.syncToLocalFs(taskFileName, tc)
+}
+
+func (t *taskContainer) Load(taskFileName string) (interface{}, bool) {
+	return t.syncTaskMap.Load(taskFileName)
+}
+
+func (t *taskContainer) Range(f func(key, value interface{}) bool) {
+	t.syncTaskMap.Range(f)
+}
+
+func (t *taskContainer) Delete(taskFileName string) {
+	t.syncTaskMap.Delete(taskFileName)
+	fp := t.ps.taskMetaFilePath(taskFileName)
+	fpBak := t.ps.taskMetaFileBakPath(taskFileName)
+
+	os.Remove(fp)
+	os.Remove(fpBak)
+}
+
+func (t *taskContainer) syncToLocalFs(key string, tc *taskConfig) {
+	fpBak := t.ps.taskMetaFileBakPath(key)
+	data, err := json.Marshal(tc)
+	if err != nil {
+		logrus.Warnf("failed to sync task %s to local fs: %v", key, err)
+		return
+	}
+
+	err = ioutil.WriteFile(fpBak, data, 0664)
+	if err != nil {
+		logrus.Warnf("failed to sync task %s to local fs: %v", key, err)
+		return
+	}
+
+	fp := t.ps.taskMetaFilePath(key)
+	err = os.Rename(fpBak, fp)
+	if err != nil {
+		logrus.Warnf("failed to sync task %s to local fs: %v", key, err)
+	}
 }
 
 // taskConfig refers to some name about peer task.
 type taskConfig struct {
-	taskID     string
-	rateLimit  int
-	cid        string
-	dataDir    string
-	superNode  string
-	finished   bool
-	accessTime time.Time
-
-	cache		*cacheBuffer
+	TaskID     string		`json:"taskID"`
+	RateLimit  int			`json:"rateLimit"`
+	Cid        string		`json:"cid"`
+	DataDir    string		`json:"dataDir"`
+	SuperNode  string		`json:"superNode"`
+	Finished   bool			`json:"finished"`
+	AccessTime time.Time	`json:"accessTime"`
+	Other      *api.FinishTaskOther `json:"other"`
+	cache		*cacheBuffer	`json:"-"`
 }
 
 // uploadParam refers to all params needed in the handler of upload.
@@ -110,8 +172,148 @@ type uploadParam struct {
 	pieceNum  int64
 }
 
+const(
+	taskMetaDir = "task-meta"
+)
+
+// reload the resource which read from local file system
+func (ps *peerServer) reload() {
+	var(
+		localTaskConfig = map[string]*taskConfig{}
+		err error
+	)
+
+	taskMetaPath := filepath.Join(ps.cfg.RV.MetaPath, taskMetaDir)
+	err = os.MkdirAll(taskMetaPath, 0744)
+	if err != nil {
+		panic(fmt.Sprintf("failed to init dir %s", taskMetaPath))
+	}
+
+	localTaskConfig, err = ps.readTaskInfoFromDir(filepath.Join(ps.cfg.RV.MetaPath, taskMetaDir))
+	if err != nil {
+		logrus.Warnf("failed to read task from local: %v", err)
+		return
+	}
+
+	logrus.Infof("try to reload task file")
+	ps.initLocalTask(localTaskConfig)
+	ps.registerTaskToSuperNode()
+}
+
+func (ps *peerServer) registerTaskToSuperNode() {
+	ps.syncTaskContainer.Range(func(key, value interface{}) bool {
+		taskFileName := key.(string)
+		tc := value.(*taskConfig)
+		ps.reportResource(taskFileName, tc)
+		return true
+	})
+}
+
+func (ps *peerServer) reportResource(taskFileName string, tc *taskConfig) {
+	req := &types.RegisterRequest{
+		RawURL: tc.Other.RawURL,
+		TaskURL: tc.Other.TaskURL,
+		Cid: ps.cfg.RV.Cid,
+		IP: ps.cfg.RV.LocalIP,
+		Port: ps.cfg.RV.PeerPort,
+		Path: taskFileName,
+		Md5: ps.cfg.Md5,
+		Identifier: ps.cfg.Identifier,
+		Headers: tc.Other.Headers,
+		Dfdaemon: ps.cfg.DFDaemon,
+		Insecure: ps.cfg.Insecure,
+		TaskId: tc.TaskID,
+		FileLength: tc.Other.FileLength,
+	}
+
+	for _,node := range ps.cfg.Nodes {
+		resp, err := ps.api.ReportResource(node, req)
+		if err == nil && resp.Code == constants.Success {
+			logrus.Infof("success to report resource %v to supernode", req)
+			break
+		}else{
+			logrus.Errorf("failed to report resource %v tp supernode, resp: %v, err: %v", req, resp, err)
+		}
+	}
+}
+
+func (ps *peerServer) readTaskInfoFromDir(dir string) (map[string]*taskConfig, error) {
+	result := map[string]*taskConfig{}
+
+	fin, err := ioutil.ReadDir(dir)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, fi := range fin {
+		taskFileName := ""
+		isBak := false
+
+		if strings.HasSuffix(fi.Name(), ".meta") {
+			taskFileName = strings.TrimSuffix(fi.Name(), ".meta")
+		}else if strings.HasSuffix(fi.Name(), ".meta.bak") {
+			taskFileName = strings.TrimSuffix(fi.Name(), ".meta.bak")
+			isBak = true
+		}else{
+			continue
+		}
+
+		fp  := filepath.Join(dir, fi.Name())
+		meta, err := ioutil.ReadFile(fp)
+		if err != nil {
+			logrus.Errorf("failed to read meta data of task %s: %v", fi.Name(), err)
+			continue
+		}
+
+		tc := &taskConfig{}
+
+		err = json.Unmarshal(meta, tc)
+		if err != nil {
+			logrus.Errorf("failed to read meta data of task %s: %v", fi.Name(), err)
+			continue
+		}
+
+		if !isBak {
+			result[taskFileName] = tc
+		}else{
+			if _, exist := result[taskFileName]; !exist {
+				result[taskFileName] = tc
+			}
+		}
+	}
+
+	return result, nil
+}
+
+func (ps *peerServer) initLocalTask(m map[string]*taskConfig) {
+	for taskFileName, tc := range m {
+		tc.AccessTime = time.Now()
+
+		taskPath := helper.GetServiceFile(taskFileName, tc.DataDir)
+
+		_, err := os.Stat(taskPath)
+		if err != nil {
+			logrus.Warnf("failed to stat taskFile %s for task %s: %v", taskPath, tc.TaskID, err)
+			continue
+		}
+
+		tc.cache = newCacheBuffer()
+		ps.syncTaskContainer.Store(taskFileName, tc)
+		ps.syncCache(taskFileName)
+		// todo: notify the supernode to register task
+	}
+}
+
+func (ps *peerServer) taskMetaFilePath(taskFileName string) string {
+	return filepath.Join(ps.cfg.RV.MetaPath, taskMetaDir, fmt.Sprintf("%s.meta", taskFileName))
+}
+
+func (ps *peerServer) taskMetaFileBakPath(taskFileName string) string {
+	return filepath.Join(ps.cfg.RV.MetaPath, taskMetaDir, fmt.Sprintf("%s.meta.bak", taskFileName))
+}
+
 // ----------------------------------------------------------------------------
-// init method of peerServer
+// reload method of peerServer
 
 func (ps *peerServer) initRouter() *mux.Router {
 	r := mux.NewRouter()
@@ -120,7 +322,7 @@ func (ps *peerServer) initRouter() *mux.Router {
 	r.HandleFunc(config.LocalHTTPPathCheck+"{commonFile:.*}", ps.checkHandler).Methods("GET")
 	r.HandleFunc(config.LocalHTTPPathClient+"finish", ps.oneFinishHandler).Methods("GET")
 	r.HandleFunc(config.LocalHTTPPing, ps.pingHandler).Methods("GET")
-
+	r.HandleFunc(config.LocalHTTPFETCH, ps.fetchHandler).Methods("GET")
 	return r
 }
 
@@ -152,7 +354,7 @@ func (ps *peerServer) uploadHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// get task config
-	v, ok := ps.syncTaskMap.Load(taskFileName)
+	v, ok := ps.syncTaskContainer.Load(taskFileName)
 	if !ok {
 		rangeErrorResponse(w, fmt.Errorf("failed to get taskPath: %s", taskFileName))
 		logrus.Errorf("failed to open file:%s, %v", taskFileName, err)
@@ -160,6 +362,9 @@ func (ps *peerServer) uploadHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	task := v.(*taskConfig)
+	// sync access time
+	task.AccessTime = time.Now()
+
 	if task.cache.valid {
 		// try to upload from cache
 		size = task.cache.size
@@ -233,15 +438,15 @@ func (ps *peerServer) parseRateHandler(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		w.WriteHeader(http.StatusBadRequest)
 		fmt.Fprint(w, err.Error())
-		logrus.Errorf("failed to convert rateLimit %v, %v", rateLimit, err)
+		logrus.Errorf("failed to convert RateLimit %v, %v", rateLimit, err)
 		return
 	}
 	sendSuccess(w)
 
-	// update the rateLimit of commonFile
-	if v, ok := ps.syncTaskMap.Load(taskFileName); ok {
+	// update the RateLimit of commonFile
+	if v, ok := ps.syncTaskContainer.Load(taskFileName); ok {
 		param := v.(*taskConfig)
-		param.rateLimit = clientRate
+		param.RateLimit = clientRate
 	}
 
 	// no need to calculate rate when totalLimitRate less than or equals zero.
@@ -278,9 +483,9 @@ func (ps *peerServer) checkHandler(w http.ResponseWriter, r *http.Request) {
 	dataDir := r.Header.Get(config.StrDataDir)
 
 	param := &taskConfig{
-		dataDir: dataDir,
+		DataDir: dataDir,
 	}
-	ps.syncTaskMap.Store(taskFileName, param)
+	ps.syncTaskContainer.Store(taskFileName, param)
 	fmt.Fprintf(w, "%s@%s", taskFileName, version.DFGetVersion)
 }
 
@@ -296,30 +501,48 @@ func (ps *peerServer) oneFinishHandler(w http.ResponseWriter, r *http.Request) {
 	taskID := r.FormValue(config.StrTaskID)
 	cid := r.FormValue(config.StrClientID)
 	superNode := r.FormValue(config.StrSuperNode)
+	other := r.FormValue(config.StrOther)
+	otherData, err := base64.StdEncoding.DecodeString(other)
+	if err != nil {
+		sendHeader(w, http.StatusBadRequest)
+		fmt.Fprintf(w, "invalid params, failed to decode other data: %v", err)
+		return
+	}
+
+	otherSt := api.FinishTaskOther{}
+	err = json.Unmarshal(otherData, &otherSt)
+	if err != nil {
+		sendHeader(w, http.StatusBadRequest)
+		fmt.Fprintf(w, "invalid params, failed to decode other data: %v", err)
+		return
+	}
+
 	if taskFileName == "" || taskID == "" || cid == "" {
 		sendHeader(w, http.StatusBadRequest)
 		fmt.Fprintf(w, "invalid params")
 		return
 	}
 
-	if v, ok := ps.syncTaskMap.Load(taskFileName); ok {
+	if v, ok := ps.syncTaskContainer.Load(taskFileName); ok {
 		task := v.(*taskConfig)
-		task.taskID = taskID
-		task.rateLimit = 0
-		task.cid = cid
-		task.superNode = superNode
-		task.finished = true
-		task.accessTime = time.Now()
+		task.TaskID = taskID
+		task.RateLimit = 0
+		task.Cid = cid
+		task.SuperNode = superNode
+		task.Finished = true
+		task.AccessTime = time.Now()
 		task.cache = newCacheBuffer()
+		task.Other = &otherSt
 	} else {
-		ps.syncTaskMap.Store(taskFileName, &taskConfig{
-			taskID:     taskID,
-			cid:        cid,
-			dataDir:    ps.cfg.RV.SystemDataDir,
-			superNode:  superNode,
-			finished:   true,
-			accessTime: time.Now(),
-			cache: newCacheBuffer(),
+		ps.syncTaskContainer.Store(taskFileName, &taskConfig{
+			TaskID:     taskID,
+			Cid:        cid,
+			DataDir:    ps.cfg.RV.SystemDataDir,
+			SuperNode:  superNode,
+			Finished:   true,
+			AccessTime: time.Now(),
+			cache:      newCacheBuffer(),
+			Other:      &otherSt,
 		})
 	}
 	go ps.syncCache(taskFileName)
@@ -328,7 +551,7 @@ func (ps *peerServer) oneFinishHandler(w http.ResponseWriter, r *http.Request) {
 }
 
 func (ps *peerServer) syncCache(taskFileName string) {
-	v, ok := ps.syncTaskMap.Load(taskFileName)
+	v, ok := ps.syncTaskContainer.Load(taskFileName)
 	if !ok {
 		return
 	}
@@ -368,7 +591,7 @@ func (ps *peerServer) pingHandler(w http.ResponseWriter, r *http.Request) {
 func (ps *peerServer) getTaskFile(taskFileName string) (*os.File, int64, error) {
 	errSize := int64(-1)
 
-	v, ok := ps.syncTaskMap.Load(taskFileName)
+	v, ok := ps.syncTaskContainer.Load(taskFileName)
 	if !ok {
 		return nil, errSize, fmt.Errorf("failed to get taskPath: %s", taskFileName)
 	}
@@ -377,10 +600,10 @@ func (ps *peerServer) getTaskFile(taskFileName string) (*os.File, int64, error) 
 		return nil, errSize, fmt.Errorf("failed to assert: %s", taskFileName)
 	}
 
-	// update the accessTime of taskFileName
-	tc.accessTime = time.Now()
+	// update the AccessTime of taskFileName
+	tc.AccessTime = time.Now()
 
-	taskPath := helper.GetServiceFile(taskFileName, tc.dataDir)
+	taskPath := helper.GetServiceFile(taskFileName, tc.DataDir)
 
 	fileInfo, err := os.Stat(taskPath)
 	if err != nil {
@@ -493,13 +716,13 @@ func (ps *peerServer) calculateRateLimit(clientRate int) int {
 	// for each key and value present in the map
 	f := func(key, value interface{}) bool {
 		if task, ok := value.(*taskConfig); ok {
-			if !task.finished {
-				total += task.rateLimit
+			if !task.Finished {
+				total += task.RateLimit
 			}
 		}
 		return true
 	}
-	ps.syncTaskMap.Range(f)
+	ps.syncTaskContainer.Range(f)
 
 	// calculate the rate limit again according to totalLimit
 	if total > ps.totalLimitRate {
@@ -546,14 +769,14 @@ func (ps *peerServer) waitForShutdown() {
 
 func (ps *peerServer) shutdown() {
 	// tell supernode this peer node is down and delete related files.
-	ps.syncTaskMap.Range(func(key, value interface{}) bool {
+	ps.syncTaskContainer.Range(func(key, value interface{}) bool {
 		task, ok := value.(*taskConfig)
-		if ok {
-			ps.api.ServiceDown(task.superNode, task.taskID, task.cid)
-			serviceFile := helper.GetServiceFile(key.(string), task.dataDir)
+		if ok && !task.Other.SpecReport {
+			ps.api.ServiceDown(task.SuperNode, task.TaskID, task.Cid)
+			serviceFile := helper.GetServiceFile(key.(string), task.DataDir)
 			os.Remove(serviceFile)
 			logrus.Infof("shutdown, remove task id:%s file:%s",
-				task.taskID, serviceFile)
+				task.TaskID, serviceFile)
 		}
 		return true
 	})
@@ -570,30 +793,34 @@ func (ps *peerServer) deleteExpiredFile(path string, info os.FileInfo,
 	expireTime time.Duration) bool {
 	taskName := helper.GetTaskName(info.Name())
 	//logrus.Infof("fileName: %s, taskName: %s", info.Name(), taskName)
-	if v, ok := ps.syncTaskMap.Load(taskName); ok {
+	if v, ok := ps.syncTaskContainer.Load(taskName); ok {
 		task, ok := v.(*taskConfig)
-		if ok && !task.finished {
+		if ok && !task.Finished {
 			return false
 		}
 
-		var lastAccessTime = task.accessTime
+		var lastAccessTime = task.AccessTime
 		// use the bigger of access time and modify time to
 		// check whether the task is expired
-		if task.accessTime.Sub(info.ModTime()) < 0 {
+		if task.AccessTime.Sub(info.ModTime()) < 0 {
 			lastAccessTime = info.ModTime()
 		}
 		// if the last access time is expireTime ago
 		if time.Since(lastAccessTime) > expireTime {
 			// ignore the gc
-			//if ok {
-			//	ps.api.ServiceDown(task.superNode, task.taskID, task.cid)
-			//}
-			//os.Remove(path)
-			//ps.syncTaskMap.Delete(taskName)
-			//return true
+			if ok {
+				if !task.Other.SpecReport {
+					ps.api.ServiceDown(task.SuperNode, task.TaskID, task.Cid)
+				}else{
+					ps.api.ReportResourceDeleted(task.SuperNode, task.TaskID, task.Cid)
+				}
+			}
+			os.Remove(path)
+			ps.syncTaskContainer.Delete(taskName)
+			return true
 		}
 	} else {
-		//os.Remove(path)
+		os.Remove(path)
 		return true
 	}
 	return false
@@ -626,4 +853,59 @@ func rangeErrorResponse(w http.ResponseWriter, err error) {
 func jsonStr(v interface{}) string {
 	b, _ := json.Marshal(v)
 	return string(b)
+}
+
+// fetchHandler fetchs the local resource
+func (ps *peerServer) fetchHandler(w http.ResponseWriter, r *http.Request) {
+	if err := r.ParseForm(); err != nil {
+		sendHeader(w, http.StatusBadRequest)
+		fmt.Fprint(w, err.Error())
+		return
+	}
+
+	// todo: add filter ?
+	result := []*apitypes.TaskFetchInfo{}
+
+	ps.syncTaskContainer.Range(func(key, value interface{}) bool {
+		taskFileName := key.(string)
+		tc := value.(*taskConfig)
+
+		taskInfo := &apitypes.TaskFetchInfo{
+			Pieces: []*apitypes.PieceInfo{
+				{
+					Path: taskFileName,
+				},
+			},
+			Task: &apitypes.TaskInfo{
+				ID: 	tc.TaskID,
+				TaskURL: tc.Other.TaskURL,
+				RawURL: tc.Other.RawURL,
+				PieceSize: int32(tc.Other.FileLength),
+				PieceTotal: 1,
+				FileLength: tc.Other.FileLength,
+				HTTPFileLength: tc.Other.FileLength,
+				// Headers: tc.Other.Headers, todo: parse header
+				Md5: ps.cfg.Md5,
+				Identifier: ps.cfg.Identifier,
+			},
+		}
+
+		result = append(result, taskInfo)
+		return true
+	})
+
+	resp := &types.FetchLocalTaskInfo{
+		Tasks: result,
+	}
+
+	respData, err := json.Marshal(resp)
+	if err != nil {
+		sendHeader(w, http.StatusInternalServerError)
+		fmt.Fprintf(w, "failed to encode %v: %v", resp, err)
+		return
+	}
+
+	sendHeader(w, http.StatusOK)
+	w.Write(respData)
+	return
 }
